\name{AutoIns}
\alias{AutoIns}
\docType{data}
\title{Claims on automobile insurance policy
}
\description{The number of claims made by \eqn{k = 9461} individuals during a single year. }
\usage{data("AutoIns")}
\format{
  A vector of length 9461.
  \describe{
    \item{\code{value}}{number of auto insurance clams}
     }
}

\source{
Efron, B. and Hastie, T. (2016), "Computer Age Statistical Inference," vol. 5, Cambridge University Press.
}

\keyword{datasets}
