\name{ChildIll}
\alias{ChildIll}
\docType{data}
\title{Frequency of child illness}
\description{ Results of a study that followed \eqn{k = 602} pre-school children in north-east Thailand from June 1982 through September 1985.  Researchers recorded the number of times a child became ill during every 2-week period.
}
\usage{data("ChildIll")}
\format{
 A vector of length \eqn{k=602}.
  \describe{
    \item{\code{value}}{number of times the \eqn{i^{th}} child became ill during the study}

  }
}

\source{Bohning, D., 2000. \emph{Computer-assisted Analysis of Mixtures and Applications: Meta-analysis, Disease Mapping, and Others} (Vol. 81). CRC press.
}
\keyword{datasets}
