\name{CorbBfly}
\alias{CorbBfly}
\docType{data}
\title{ Corbet's Butterfly data
}
\description{
The number of times Alexander Corbet captured a species of butterfly during a two-year period in Malaysia.
}
\usage{data("CorbBfly")}
\format{
  A vector of length \eqn{k = 501}.
  \describe{
    \item{\code{value}}{number of times Corbet captured the \eqn{i^{th}} species}
    
  }
}
\source{
Fisher, R.A., Corbet, A.S. and Williams, C.B., 1943. "The relation between the number of species and the number of individuals in a random sample of an animal population." \emph{The Journal of Animal Ecology}, pp.42-58.}
\references{
Efron, B. and Hastie, T., 2016. \emph{Computer Age Statistical Inference} (Vol. 5). Cambridge University Press.
}
\keyword{datasets}
