\name{steroid}
\alias{steroid}
\docType{data}
\title{
Nasal Steroid Data
}
\description{
The standardized mean difference \eqn{y_i} and standard errors \eqn{se_i} for seven randomised studies on the use of topical steroids in treatment of chronic rhinosinusitis with nasal polyps.
}
\usage{data("steroid")}
\format{
  A data frame of \eqn{(y_i, se_i)} for \eqn{i = 1,...,7}.
  \describe{
    \item{\code{y}}{standard mean difference of clinical trials for topical steroids found in the \eqn{i^{th}} study}
    \item{\code{se}}{standard error of the standard mean difference for the \eqn{i^{th}} study}
  }
}
\source{ IntHout, J., Ioannidis, J. P., Rovers, M. M., & Goeman, J. J., 2016. "Plea for routinely presenting prediction intervals in meta-analysis," \emph{BMJ open, 6}(7), e010247.}

\keyword{datasets}
