\name{arsenic}
\alias{arsenic}
\docType{data}
\title{ Arsenic levels in oyster tissue
}
\description{Results from an inter-laboratory study involving \eqn{k = 28} measurements for the level of arsenic in oyster tissue. \code{y} is the mean level of arsenic from a lab and \code{se} is the standard error of the measurement.  
}
\usage{data("arsenic")}
\format{
  A data frame of \eqn{(y_i, se_i)} for \eqn{i = 1,...,28}.
  \describe{
    \item{\code{y}}{mean level of arsenic in the tissue measured by a single lab }
    \item{\code{se}}{the standard error of the measurement by a single lab}
  }
}

\source{ 
Wille, S. and Berman, S. (1995), "Ninth round intercomparison for trace metals in marine sediments and biological tissues," NRC/NOAA.
}
\references{
Possolo, A. (2013), "Five examples of assessment and expression of measurement uncertainty," Applied Stochastic Models in Business and Industry, 29, 1-18.
}
\keyword{datasets}
