\name{AutoIns}
\alias{AutoIns}
\docType{data}
\title{Number of claims on an insurance policy}

\description{The number of claims on an automobile insurance policy made by \eqn{k = 9461} individuals during a single year. }
\usage{data("AutoIns")}
\format{
  A vector of length 9461.
  \describe{
    \item{\code{value}}{number of auto insurance claims by the \eqn{i^{th}} person}
     }
}

\source{
Efron, B. and Hastie, T., 2016. \emph{Computer Age Statistical Inference} (Vol. 5). Cambridge University Press.
}

\keyword{datasets}
