\name{BayesGOF-package}
\alias{BayesGOF-package}
\alias{BayesGOF}
\docType{package}
\title{
\packageTitle{BayesGOF}
}
\description{
A Bayesian data modeling scheme that performs four interconnected tasks: (i) characterizes the uncertainty of the elicited parametric prior; (ii) provides exploratory diagnostic for checking prior-data conflict; (iii) computes the final statistical prior density estimate; and (iv) executes macro- and micro-inference.  
}

\references{
Mukhopadhyay, S. and Fletcher, D. (2018), "Bayesian Modeling via Goodness of Fit," Technical report, \url{https://arxiv.org/abs/1802.00474 }.
}
\keyword{ package }

