\name{BayesGOF-package}
\alias{BayesGOF-package}
\alias{BayesGOF}
\docType{package}
\title{
\packageTitle{BayesGOF}
}
\description{
Nonparametric method for learning prior distribution starting with parametric (subjective) prior \eqn{g}.  It performs four interconnected tasks: (i) characterizes the uncertainity of the a priori elicited \eqn{g}; (ii) exploratory diagnostic for checking prior-data conflict; (iii) computes the final statistical prior density estimate \eqn{\hat{\pi}(\theta)}; and (iv) performs macro- and micro-inference.  
}

\references{
Mukhopadhyay, S. and Fletcher, D. (2017), "Bayes Modeling via Goodness of Fit," Technical report.
}
\keyword{ package }

