\name{ChildIll}
\alias{ChildIll}
\docType{data}
\title{ Child illness data
}
\description{ Results of a study where researchers followed \eqn{k = 602} pre-school children in north-east Thailand from June 1982 through September 1985.  Researchers recorded the number of times a child became ill during every 2-week period.
}
\usage{data("ChildIll")}
\format{
 A vector of length \eqn{k=602}.
  \describe{
    \item{\code{value}}{number of times child \eqn{i} became ill during the study.}

  }
}

\source{Bohning, D. (2000), "Computer-assisted Analysis of Mixtures and Applications: Meta-analysis, Disease Mapping, and Others," Boca Raton: Chapman and Hall-CRC.
}
\keyword{datasets}
